Update slack_event_handler.py
In this updated version, we ensure that the OAuth flow retrieves the user token (xoxp-) rather than the bot token (xoxb-). If the access_token field does not exist in the authed_user dictionary, the code will now log an error and terminate the OAuth flow.

After deploying this version, the app should store and use the user token correctly.